staging/pkg/mod/golang.org/x/net@v0.0.0-20210520170846-37e1c6afe023/html/testdata/go/issue_30600_parse_panics_in_cell_mode.dat

#data
<table><math><th><mo><select></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math th>
|         <math mo>
|           <select>
|     <table>
